pyproject.toml
[tool.poetry]
name = "european-comission-network"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
bs4 = "^0.0.2"
networkx = "^3.4.2"
selenium = "^4.27.1"
webdriver-manager = "^4.0.2"
langchain-community = "^0.3.14"
pypdf = "^5.1.0"
langchain-openai = "^0.2.14"
tavily-python = "^0.5.0"
faiss-cpu = "^1.9.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
